[JENKINS-46064] CamelCase as requested
