+++ pipeline-model-definition/src/main/java/org/jenkinsci/plugins/pipeline/modeldefinition/when/impl/ChangeLogConditional.java
@@ -33,42 +33,39 @@ import org.jenkinsci.plugins.pipeline.modeldefinition.parser.ASTParserUtils;
 import org.jenkinsci.plugins.pipeline.modeldefinition.when.DeclarativeStageConditional;
 import org.jenkinsci.plugins.pipeline.modeldefinition.when.DeclarativeStageConditionalDescriptor;
 import org.kohsuke.stapler.DataBoundConstructor;
-import org.kohsuke.stapler.DataBoundSetter;
 
 import javax.annotation.CheckForNull;
+import java.util.regex.Pattern;
 
 /**
- * Conditional that checks the affected file paths recorded in the changelog.
+ * Conditional that checks the messages in the changelog.
  *
  * The build must first have collected the changelog via for example <code>checkout scm</code>.
  */
-public class ChangesetConditional extends DeclarativeStageConditional<ChangesetConditional> {
+public class ChangeLogConditional extends DeclarativeStageConditional<ChangeLogConditional> {
 
-    private String glob;
-    private boolean caseSensitive;
+    private Pattern pattern;
+    private Pattern multiLinePattern;
 
     @DataBoundConstructor
-    public ChangesetConditional(String glob) {
-        this.glob = glob;
-        this.caseSensitive = false;
+    public ChangeLogConditional(String pattern) {
+        //TODO JENKINS-46065 validate the regexp when #179 is merged
+        this.pattern = Pattern.compile(pattern);
+        this.multiLinePattern = Pattern.compile("(?m)(?s)^[^\\r\\n]*?" + pattern + "[^\\r\\n]*?$",
+                Pattern.MULTILINE | Pattern.DOTALL);
     }
 
-    public String getGlob() {
-        return glob;
+    public Pattern getPattern() {
+        return pattern;
     }
 
-    public boolean isCaseSensitive() {
-        return caseSensitive;
-    }
-
-    @DataBoundSetter
-    public void setCaseSensitive(boolean caseSensitive) {
-        this.caseSensitive = caseSensitive;
+    public Pattern getMultiLinePattern() {
+        return multiLinePattern;
     }
 
     @Extension
-    @Symbol("changeset")
-    public static class DescriptorImpl extends DeclarativeStageConditionalDescriptor<ChangesetConditional> {
+    @Symbol("changelog")
+    public static class DescriptorImpl extends DeclarativeStageConditionalDescriptor<ChangeLogConditional> {
         @Override
         public Expression transformToRuntimeAST(@CheckForNull ModelASTWhenContent original) {
             return ASTParserUtils.transformWhenContentToRuntimeAST(original);
